+++ web3walletcore/src/commonMain/kotlin/com/sonsofcrypto/web3walletcore/modules/mnemonicConfirmation/MnemonicConfirmationInteractor.kt
@@ -1,6 +1,7 @@
 package com.sonsofcrypto.web3walletcore.modules.mnemonicConfirmation
 
 import com.sonsofcrypto.web3lib.services.keyStore.SignerStoreService
+import com.sonsofcrypto.web3walletcore.common.helpers.MnemonicPresenterHelper
 import com.sonsofcrypto.web3walletcore.services.actions.Action
 import com.sonsofcrypto.web3walletcore.services.actions.ActionsService
 import com.sonsofcrypto.web3walletcore.services.mnemonic.MnemonicService
@@ -8,14 +9,18 @@ import com.sonsofcrypto.web3walletcore.services.mnemonic.MnemonicServiceError
 import com.sonsofcrypto.web3walletcore.services.mnemonic.MnemonicWord
 
 interface MnemonicConfirmationInteractor {
-    fun prefix(mnemonic: String, cursorLocation: Int): String
-    fun potentialMnemonicWords(prefix: String?): List<String>
-    fun findInvalidWords(mnemonic: String?): List<MnemonicWord>
-    fun isValidPrefix(prefix: String): Boolean
-    fun isMnemonicValid(mnemonic: String, salt: String?): Boolean
+    var mnemonicStr: String
+    var cursorLoc: Int
+    var salt: String
+
+    fun mnemonic(): String
+    fun potentialMnemonicWords(): List<String>
+    fun mnemonicWordsInfo(): List<MnemonicWord>
+    fun isMnemonicValid(): Boolean
+    fun mnemonicError(): MnemonicServiceError?
+
     fun showSalt(): Boolean
     fun markConfirmActionComplete()
-    fun mnemonicError(mnemonic: List<String>, salt: String): MnemonicServiceError?
 }
 
 class DefaultMnemonicConfirmationInteractor(
@@ -23,21 +28,31 @@ class DefaultMnemonicConfirmationInteractor(
     private val actionsService: ActionsService,
     private val mnemonicService: MnemonicService,
 ): MnemonicConfirmationInteractor {
+    private val helper = MnemonicPresenterHelper()
 
-    override fun prefix(mnemonic: String, cursorLocation: Int): String =
-        mnemonicService.prefix(mnemonic, cursorLocation)
+    override var mnemonicStr: String = ""
+    override var cursorLoc: Int = 0
+    override var salt: String = ""
 
-    override fun potentialMnemonicWords(prefix: String?): List<String> =
-        mnemonicService.potentialMnemonicWords(prefix)
+    override fun mnemonic(): String =
+        mnemonicStr.trim()
 
-    override fun findInvalidWords(mnemonic: String?): List<MnemonicWord> =
-        mnemonicService.findInvalidWords(mnemonic)
+    override fun potentialMnemonicWords(): List<String> = mnemonicService
+        .potentialMnemonicWords(mnemonicService.prefix(mnemonicStr, cursorLoc))
 
-    override fun isValidPrefix(prefix: String): Boolean =
-        mnemonicService.isValidPrefix(prefix)
+    override fun mnemonicWordsInfo(): List<MnemonicWord> {
+        val prefix = mnemonicService.prefix(mnemonicStr, cursorLoc)
+        val words = mnemonicService.findInvalidWords(mnemonicStr)
+        return helper.updateWordsInfo(words, prefix, cursorLoc) {
+            mnemonicService.isValidPrefix(it)
+        }
+    }
 
-    override fun isMnemonicValid(mnemonic: String, salt: String?): Boolean =
-        mnemonicService.mnemonicError(mnemonic, salt) == null
+    override fun isMnemonicValid(): Boolean =
+        mnemonicError() == null
+
+    override fun mnemonicError(): MnemonicServiceError? =
+        mnemonicService.mnemonicError(mnemonicStr, salt)
 
     override fun showSalt(): Boolean {
         val keyStoreItem = signerStoreService.selected ?: return false
@@ -47,10 +62,4 @@ class DefaultMnemonicConfirmationInteractor(
     override fun markConfirmActionComplete() {
         actionsService.markComplete(Action.ConfirmMnemonic)
     }
-
-    override fun mnemonicError(
-        mnemonic: List<String>,
-        salt: String
-    ): MnemonicServiceError? =
-        mnemonicService.mnemonicError(mnemonic.joinToString(" "), salt)
 }